# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Bingo Builder API
  description: >-
    RESTful API for creating, managing, and playing multiplayer bingo cards.


    **Security Features:**

    - CORS protection with origin restriction

    - Multi-tier rate limiting (API, write, peer, list operations)

    - Request validation via OpenAPI specification (auto-generated from Zod
    schemas)

    - Request body size limits (1MB)


    **Note:** This OpenAPI specification is automatically generated from Zod
    schemas.

    Do not edit this file manually. Instead, update the schemas in
    `backend/schemas/` and regenerate.
  version: 1.0.0
  contact:
    name: Bingo Builder
servers:
  - url: http://localhost:3001
    description: Development server
  - url: http://localhost:3002
    description: Test server
...
components:
  schemas:
    Card:
      type: object
      properties:
        _id:
          type: string
          pattern: ^[0-9a-fA-F]{24}$
          description: MongoDB ObjectId
          example: 507f1f77bcf86cd799439011
        title:
          type: string
          maxLength: 25
          description: Card title
        createdBy:
          type: string
          default: ""
          description: Creator name
        rows:
          type: integer
          minimum: 2
          maximum: 5
          description: Number of rows
        columns:
          type: integer
          minimum: 2
          maximum: 6
          description: Number of columns
        tiles:
          type: array
          items:
            $ref: "#/components/schemas/Tile"
          description: Array of tiles
        isPublished:
          type: boolean
          default: false
          description: Whether the card is published and playable
        publishedAt:
          type: string
          nullable: true
          format: date-time
          description: When the card was published
        createdAt:
          type: string
          format: date-time
          description: When the card was created
        updatedAt:
          type: string
          format: date-time
          description: When the card was last updated
      required:
        - _id
        - title
        - rows
        - columns
        - tiles
        - createdAt
        - updatedAt
    Tile:
      type: object
      properties:
        value:
          type: string
          maxLength: 40
          default: ""
          description: Content of the tile
          example: Free space
        position:
          type: integer
          minimum: 0
          description: Position in the grid (0-indexed)
          example: 0
      required:
        - position
    CardInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 25
          description: Card title
          example: My Bingo Card
        createdBy:
          type: string
          default: ""
          description: Creator name (optional)
          example: John Doe
        rows:
          type: integer
          minimum: 2
          maximum: 5
          description: Number of rows in the grid
          example: 3
        columns:
          type: integer
          minimum: 2
          maximum: 6
          description: Number of columns in the grid
          example: 3
        tiles:
          type: array
          items:
            $ref: "#/components/schemas/Tile"
          minItems: 4
          maxItems: 30
          description: Array of tiles (length must equal rows × columns)
      required:
        - title
        - rows
        - columns
        - tiles
    CardUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 25
        createdBy:
          type: string
        rows:
          type: integer
          minimum: 2
          maximum: 5
        columns:
          type: integer
          minimum: 2
          maximum: 6
        tiles:
          type: array
          items:
            $ref: "#/components/schemas/Tile"
          minItems: 4
          maxItems: 30
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: Card not found
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
              errorCode:
                type: string
            required:
              - message
          description: Detailed validation errors (for 400 responses)
      required:
        - message
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          description: Success message
      required:
        - message
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
        deletedCount:
          type: integer
          description: Number of deleted items
      required:
        - message
        - deletedCount
    UpdateResponse:
      type: object
      properties:
        message:
          type: string
        modifiedCount:
          type: integer
          description: Number of modified items
      required:
        - message
        - modifiedCount
...
paths:
  /api/cards:
    get:
      summary: List all cards
      description: Retrieve all bingo cards sorted by creation date (newest first).
        Rate limited to prevent scraping.
      tags:
        - cards
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Card"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new card
      description: Create a new bingo card with specified grid size and tiles.
        Stricter rate limits apply.
      tags:
        - cards
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardInput"
      responses:
        "201":
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cards/{id}:
    get:
      summary: Get a single card
      description: Retrieve a specific bingo card by its ID
      tags:
        - cards
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId
          name: id
          in: path
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a card
      description: Update an unpublished card. Published cards cannot be edited.
      tags:
        - cards
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardUpdate"
      responses:
        "200":
          description: Card updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access forbidden (not the owner or card is published)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a card
      description: Delete an unpublished card. Published cards cannot be deleted.
      tags:
        - cards
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId
          name: id
          in: path
        - schema:
            type: string
            description: Creator name for ownership verification
          required: false
          description: Creator name for ownership verification
          name: createdBy
          in: query
      responses:
        "200":
          description: Card deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "403":
          description: Access forbidden (not the owner or card is published)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cards/{id}/publish:
    post:
      summary: Publish a card
      description: Publish a card to make it playable. All tiles must be filled.
      tags:
        - cards
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                createdBy:
                  type: string
                  description: Creator name for ownership verification
      responses:
        "200":
          description: Card published successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access forbidden (not the owner or card is published)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cards/{id}/unpublish:
    post:
      summary: Unpublish a card
      description: Unpublish a card to make it editable again
      tags:
        - cards
      parameters:
        - schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
            description: MongoDB ObjectId
            example: 507f1f77bcf86cd799439011
          required: true
          description: MongoDB ObjectId
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                createdBy:
                  type: string
                  description: Creator name for ownership verification
      responses:
        "200":
          description: Card unpublished successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access forbidden (not the owner or card is published)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cards/delete-by-creator:
    post:
      summary: Delete all cards by creator
      description: Delete all cards created by a specific user
      tags:
        - cards
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                createdBy:
                  type: string
                  minLength: 1
                  description: Creator name
              required:
                - createdBy
      responses:
        "200":
          description: Cards deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/cards/update-creator:
    post:
      summary: Update creator name
      description: Update creator name for all cards with the old name
      tags:
        - cards
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                oldName:
                  type: string
                  minLength: 1
                  description: Current creator name
                newName:
                  type: string
                  minLength: 1
                  description: New creator name
              required:
                - oldName
                - newName
      responses:
        "200":
          description: Creator name updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "400":
          description: Invalid request parameters or validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          headers:
            RateLimit-Limit:
              schema:
                type: integer
                description: Maximum requests allowed
              required: true
              description: Maximum requests allowed
            RateLimit-Remaining:
              schema:
                type: integer
                description: Requests remaining
              required: true
              description: Requests remaining
            RateLimit-Reset:
              schema:
                type: integer
                description: Time when limit resets (Unix timestamp)
              required: true
              description: Time when limit resets (Unix timestamp)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...